Update TDD added

+++ userAuthentication/src/main/java/com/auth/userAuthentication/data/repository/UserRepository.java
@@ -5,10 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     User findByUsername(String username);
 
     User findByEmail(String email);
-
-//    Optional<User> findByEmail(String email);
 }
